Move space state logic out of HID to new module and class.
